Switch ctrl and option
I was not able to live with ctrl set to the right thumb cluster area.
Actually as previoulsy my left hand wants in charge for ctrl and option
though the urge is much higher for ctrl. Therefore setting option
(previously not shihfted) to the right thumb cluster and ctrl back to
the left.